requirements.txt
networkx==2.8
numpy==1.20.1
Pillow==8.2.0
scikit_learn==1.0.2
scipy==1.6.2
dgl==1.0.2
pytorch==1.10.0
